datasets/characters/requirements.txt
    docopt==0.6.2
Flask==0.12.2
ImageHash==3.4
jinja2==2.10
more-itertools==2.2
Pillow==4.3.0
psutil==5.4.2
pymongo>=3.6.0
python-magic==0.4.15
termcolor==1.1.0
Werkzeug==0.12.2
Flask-Cors==3.0.3